change types to double

+++ main/modules/motor.h
@@ -11,8 +11,8 @@ class Motor {
 public:
     virtual bool is_running() = 0;
     virtual void stop() = 0;
-    virtual int32_t position() = 0;
-    virtual void position(const int32_t position, const int32_t speed, const uint32_t acceleration) = 0;
-    virtual int32_t speed() = 0;
-    virtual void speed(const int32_t speed, const uint32_t acceleration) = 0;
+    virtual double position() = 0;
+    virtual void position(const double position, const double speed, const uint32_t acceleration) = 0;
+    virtual double speed() = 0;
+    virtual void speed(const double speed, const uint32_t acceleration) = 0;
 };
